requirements/base.txt
# Base application requirements

django>=4.0.0,<5.0.0
django-ninja
django-cors-headers
pyjwt
pytz
llama-cpp-python==0.1.55
langchain==0.0.183
django-instant